openssl: Fix for the EAP-FAST on the latest OpenSSL.
CVE-2014-0224 code added a logic to reject the ChangeCipherSpec
message if it is received in incorrect order.
Normally the TLS client sends the Finished message before the
ChangeCipherSpec message is received from the server.
The code now appears to set a flag SSL3_FLAGS_CCS_OK
when Finished is sent by the client. And the code checks this
flag when the ChangeCipherSpec message is received; rejecting the
CCS packet if the flag is not set.
For TLS session resumption, the server sends the ChangeCipherSpec
before the client sends the Finished message.
This is resulting in the ChangeCipherSpec packet being rejected
for EAP-FAST session resumption because the SSL3_FLAGS_CCS_OK is
not set.

Change-Id: I5c21ed18387414ce62fd30fc78c3fccf0b8598b6
CRs-Fixed: 757998